Rewrite in Vite and TypeScript
This is a major refactoring. Vite is the bundler for vue. Although, the
individual components are not migrated the structure and build process
is ready.
