+++ .gitignore
@@ -1,9 +1,34 @@
-.DS_Store
 .serverless/
-static/data/
+public/static/data/
 dump/
 db.sqlite3
 config/prod.json
 config/local.json
 config/dev.json
-node_modules/
+
+# Logs
+logs
+*.log
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+pnpm-debug.log*
+lerna-debug.log*
+
+node_modules
+dist
+dist-ssr
+*.local
+.env.*
+
+
+# Editor directories and files
+.vscode/*
+!.vscode/extensions.json
+.idea
+.DS_Store
+*.suo
+*.ntvs*
+*.njsproj
+*.sln
+*.sw?
