+++ config/sample.json
@@ -2,5 +2,6 @@
     "OAuthClientID": "",
     "OAuthClientSecret": "",
     "OAuthRedirectURI": "",
-    "Domain": ""
+    "FrontendDomain": "",
+    "BackendDomain": ""
 }
