+++ handlers/logout.js
@@ -10,7 +10,7 @@ module.exports.handler = (event, context, callback) => {
 
   const clientKey = process.env.OAUTH_CLIENT_ID;
   const clientSecret = process.env.OAUTH_CLIENT_SECRET;
-  const domain = process.env.DOMAIN;
+  const domain = process.env.FRONTEND_DOMAIN;
 
   request(
     {
