+++ handlers/static.js
@@ -1,7 +1,7 @@
 const StaticFileHandler = require("serverless-aws-static-file-handler");
 const path = require("path");
 
-const clientFilesPath = path.join(__dirname, "./../static/");
+const clientFilesPath = path.join(__dirname, "./../dist/");
 const fileHandler = new StaticFileHandler(clientFilesPath);
 
 module.exports.handler = async (event, context) => {
@@ -10,6 +10,10 @@ module.exports.handler = async (event, context) => {
   if (event.path === "/" || event.path === "") {
     event.path = "index.html";
   }
+
+  // I don't know why, but the default behavior is to only use the last path
+  // component unless we intefere with this.
+  event.pathParameters = null;
 
   return fileHandler.get(event, context);
 };
