+++ package.json
@@ -2,11 +2,26 @@
   "name": "chartbrainz",
   "version": "1.0.0",
   "description": "",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vue-tsc && vite build",
+    "preview": "vite preview"
+  },
   "dependencies": {
+    "jquery": "^3.6.1",
+    "vue": "^3.2.41",
     "express": "^4.17.1",
     "request": "^2.88.2",
     "serverless-aws-static-file-handler": "^4.0.1",
     "serverless-offline": "^11.5.0",
     "tough-cookie": "^4.1.2"
+  },
+  "devDependencies": {
+    "@types/jquery": "^3.5.14",
+    "@vitejs/plugin-vue": "^3.2.0",
+    "typescript": "^4.6.4",
+    "vite": "^3.2.3",
+    "vue-tsc": "^1.0.9"
   }
 }
