--- static/index.html
@@ -1,216 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta
-      name="viewport"
-      content="width=device-width, initial-scale=1, shrink-to-fit=no"
-    />
-    <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
-
-    <title>ChartBrainz</title>
-
-    <link
-      type="text/css"
-      rel="stylesheet"
-      href="https://unpkg.com/bootstrap/dist/css/bootstrap.min.css"
-    />
-    <link
-      rel="stylesheet"
-      href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css"
-    />
-
-    <script src="https://unpkg.com/vue@latest/dist/vue.global.js"></script>
-    <script
-      src="https://code.jquery.com/jquery-3.6.1.min.js"
-      integrity="sha256-o88AwQnZB+VDvE9tvIXrMQaPlFFSUTR+nldQm1LuPXQ="
-      crossorigin="anonymous"
-    ></script>
-    <script
-      src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/2.9.2/umd/popper.min.js"
-      integrity="sha512-2rNj2KJ+D8s1ceNasTIex6z4HWyOnEYLVC3FigGOmyQCZc2eBXKgOxQmo3oKLHyfcj53uz4QMsRCWNbLd32Q1g=="
-      crossorigin="anonymous"
-      referrerpolicy="no-referrer"
-    ></script>
-    <script
-      src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/js/bootstrap.min.js"
-      integrity="sha384-IDwe1+LCz02ROU9k972gdyvl+AESN10+x7tBKgc9I5HFtuNz0wWnPclzo6p9vxnk"
-      crossorigin="anonymous"
-    ></script>
-
-    <style>
-      .muted_tag {
-        color: gray;
-        text-decoration: none;
-      }
-
-      .genre_tag {
-        text-decoration: none;
-      }
-
-      ul.genre {
-        margin-left: 1em;
-        padding: 0;
-        list-style: none;
-      }
-    </style>
-  </head>
-  <body>
-    <div id="app">
-      <div class="container">
-        <div class="row">
-          <div class="col-3">
-            <h1>ChartBrainz</h1>
-          </div>
-          <div class="col-9">
-            <div style="text-align: right; margin-top: 10px">
-              <input
-                type="text"
-                style="width: 200px; border: solid 1px gray"
-                v-model="search"
-                @keyup.enter="doSearch"
-              />
-              &nbsp;
-              <button type="button" class="btn btn-primary" @click="doSearch">
-                Search
-              </button>
-              &nbsp;
-              <a
-                class="btn btn-primary"
-                href="https://github.com/elliotchance/chartbrainz/issues"
-                target="_blank"
-              >
-                Report Bug / Feature Request
-              </a>
-              &nbsp;
-              <div class="btn-group" role="group" v-if="user">
-                <button type="button" class="btn btn-outline-primary">
-                  {{ ratingCount }} ratings
-                </button>
-                <button
-                  type="button"
-                  class="btn btn-primary"
-                  @click="() => syncRatings(this.user)"
-                >
-                  {{ syncStatus }}
-                </button>
-              </div>
-              &nbsp;
-              <a class="btn btn-primary" href="/login" v-if="!user"> Login </a>
-              <span v-if="user"> Hi, {{ user }} </span>
-              &nbsp;
-              <a class="btn btn-primary" v-if="user" href="/logout"> Logout </a>
-            </div>
-          </div>
-        </div>
-        <div class="row">
-          <div class="col">&nbsp;</div>
-        </div>
-        <div class="row" v-if="view === 'search'">
-          <div class="col-6">
-            <card title="Artists" title-class="h5" :close="closeSearch">
-              <span v-if="artistSearchResult.length === 0">No artists.</span>
-              <ul
-                class="list-group list-group-flush"
-                v-if="artistSearchResult.length > 0"
-              >
-                <li
-                  class="list-group-item"
-                  v-for="artist in artistSearchResult"
-                >
-                  <span
-                    class="float-end"
-                    @click="() => doArtistSearch(artist.id)"
-                  >
-                    <span class="link-primary">search</span>
-                    <chevron-right />
-                  </span>
-                  <artist
-                    :name="artist.name"
-                    :disambiguation="artist.disambiguation"
-                  />
-                </li>
-              </ul>
-            </card>
-          </div>
-          <div class="col-6">
-            <release-search-results :release-search-result="releaseSearchResult" :submit-rating="submitRating"/>
-          </div>
-        </div>
-        <div class="row" v-if="view === 'chart'">
-          <div class="col-3">
-            <card title="Filter" title-class="h5">
-              <div class="container">
-                <br />
-                <ul class="nav nav-pills">
-                  <li class="nav-item">
-                    <a
-                      :class="'nav-link ' + (decade === '' ? 'active' : '')"
-                      href="#"
-                      @click="() => setDecade('')"
-                      >All-time</a
-                    >
-                  </li>
-                  <li class="nav-item" v-for="d in availableDecades">
-                    <a
-                      :class="'nav-link ' + (decade === d ? 'active' : '')"
-                      href="#"
-                      @click="() => setDecade(d)"
-                      >{{ d }}</a
-                    >
-                  </li>
-                </ul>
-                <br />
-                <genre-selector
-                  :genres="genres"
-                  :root-genres="rootGenres"
-                  :set-current-genre="setCurrentGenre"
-                  :depth="0"
-                  :path="path"
-                  :current-genre="currentGenre"
-                  :set-path="setPath"
-                />
-              </div>
-            </card>
-          </div>
-
-          <div class="col-9">
-            <card
-              :title="`Top ${currentGenre} releases of ${decade ? 'the ' + decade : 'all-time'}`"
-              title-class="h5"
-            >
-              <h2 v-if="releases.length === 0">No releases.</h2>
-              <table class="table" v-if="releases.length > 0">
-                <thead>
-                  <tr>
-                    <th width="100">&nbsp;</th>
-                    <th width="50">&nbsp;</th>
-                    <th>&nbsp;</th>
-                    <th>&nbsp;</th>
-                    <th style="text-align: center">Average</th>
-                    <th style="text-align: center">Ratings</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  <template v-for="release, i in releases">
-                    <release-row
-                      :release="release"
-                      :number="i + 1"
-                      :submit-rating="submitRating"
-                      :is-logged-in="user !== ''"
-                    />
-                  </template>
-                </tbody>
-              </table>
-            </card>
-          </div>
-        </div>
-      </div>
-    </div>
-
-    <script src="./static/app.js"></script>
-    <script>
-      app.mount("#app");
-    </script>
-  </body>
-</html>
